backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.0
httpx==0.27.2
pydantic[email]==2.9.2
python-multipart==0.0.17

# Authentication
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0

# PostgreSQL Database (Neon)
psycopg==3.2.3
